pyproject.toml
[project]
name = "dsait-4305-polymer-prediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "transformers>=4.21.0",
    "tokenizers>=0.13.0",
    "dill>=0.4.0",
    "hydra-core>=1.3.2",
    "ipykernel>=6.29.5",
    "lightning>=2.3.3",
    "matplotlib>=3.7.5",
    "networkx>=3.1",
    "pandas>=2.0.3",
    "petname>=2.6",
    "rdkit>=2025.9.1",
    "scikit-learn>=1.3.2",
    "torch",
    "torch-geometric>=2.6.1",
    "tqdm>=4.67.1",
    "wandb>=0.22.2",
    "huggingface-hub>=0.36.0",
    "numpy>=1.24.4",
    "pytz>=2025.2",
    "python-dateutil>=2.9.0.post0",
    "egnn-pytorch>=0.2.8",
    "xgboost>=2.1.4",
]


[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
